ERP_Project/requirements.txt
djangorestframework~=3.15.1
Django~=5.0.4
django-filter~=24.2